refactor product catalogue
